# api/controllers/car.js
'use strict';
// Include our "db"
var db = require('../../config/db')();
// Exports all the functions to perform on the db
module.exports = {
    getAll,
    save,
    getOne,
    update,
    delMovie
};

// GET /movie operationId
function getAll(req, res, next) {
    res.json({
        cars: db.find()
    });
}
// POST /movie operationId
function save(req, res, next) {
    res.json({
        success: db.save(req.body),
        description: "car added to the list!"
    });
}
// GET /movie/{id} operationId
function getOne(req, res, next) {
    var id = req.swagger.params.id.value; // req.swagger contains the path parameters
    var car = db.find(id);
    if (car) {
        res.json(car);
    } else {
        res.status(204).send();
    }
}
// PUT /movie/{id} operationId
function update(req, res, next) {
    var id = req.swagger.params.id.value; // req.swagger contains the path parameters
    var car = req.body;
    if (db.update(id, car)) {
        res.json({
            success: 1,
            description: "car updated!"
        });
    } else {
        res.status(204).send();
    }

}
// DELETE /movie/{id} operationId
function delMovie(req, res, next) {
    var id = req.swagger.params.id.value; // req.swagger contains the path parameters
    if (db.remove(id)) {
        res.json({
            success: 1,
            description: "car deleted!"
        });
    } else {
        res.status(204).send();
    }

}