# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: MovieListApp
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /car/{id}:
    # our controller name
    x-swagger-router-controller: car
    get:
      operationId: getOne
      description: get a car
      # define the type of response for Success "200" and Error
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetCarResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: update
      description: update a car
      # define the parameters
      parameters:
        - name: id
          description: Car id
          type: string
          in: path
          required: true
        - name: title
          description: Car properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Car"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delMovie
      description: delete a car
      # define the parameters
      parameters:
        - name: id
          description: Car id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GetCarResponse:
    required:
      - model
      - owner
    properties:
      model:
        type: string
      owner:
        type: string
  Car:
    type: object
    properties:
      model:
        type: string
        description: task object name
      owner:
        type: string
        description: task description
    required:
      - model
      - owner
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string